Moved commands to separate system
