+++ src/main/java/org/terasology/hunger/HungerAuthoritySystem.java
@@ -34,13 +34,18 @@ import org.terasology.hunger.component.HungerComponent;
 import org.terasology.hunger.event.FoodConsumedEvent;
 import org.terasology.logic.characters.AliveCharacterComponent;
 import org.terasology.logic.common.ActivateEvent;
+import org.terasology.logic.console.commandSystem.annotations.Command;
+import org.terasology.logic.console.commandSystem.annotations.CommandParam;
+import org.terasology.logic.console.commandSystem.annotations.Sender;
 import org.terasology.logic.health.BeforeHealEvent;
 import org.terasology.logic.health.DoDamageEvent;
 import org.terasology.logic.inventory.InventoryManager;
 import org.terasology.logic.inventory.InventoryUtils;
 import org.terasology.logic.inventory.ItemComponent;
+import org.terasology.logic.permission.PermissionManager;
 import org.terasology.logic.players.event.OnPlayerRespawnedEvent;
 import org.terasology.logic.players.event.OnPlayerSpawnedEvent;
+import org.terasology.network.ClientComponent;
 import org.terasology.registry.In;
 
 /**
@@ -71,7 +76,9 @@ public class HungerAuthoritySystem extends BaseComponentSystem implements Update
     @In
     private PrefabManager prefabManager;
 
-    /** Reference to the current time, used for calculating if the food of an entity has to be decreased.*/
+    /**
+     * Reference to the current time, used for calculating if the food of an entity has to be decreased.
+     */
     @In
     private Time time;
 
@@ -164,9 +171,10 @@ public class HungerAuthoritySystem extends BaseComponentSystem implements Update
 
     /**
      * This method registers it when an entity consumes food and adds the food to the entities HungerComponent.
+     *
      * @param event The ActivateEvent called when an entity consumes food.
-     * @param item The entity which is consuming the food.
-     * @param food The Foodcomponent containing data about how much a certain type of food is filling.
+     * @param item  The entity which is consuming the food.
+     * @param food  The Foodcomponent containing data about how much a certain type of food is filling.
      */
     @ReceiveEvent
     public void foodConsumed(ActivateEvent event, EntityRef item, FoodComponent food) {
@@ -188,8 +196,9 @@ public class HungerAuthoritySystem extends BaseComponentSystem implements Update
 
     /**
      * This method deals with removal of food item after it is consumed.
+     *
      * @param event The FoodConsumedEvent called when an entity consumes food.
-     * @param item The entity which is consuming the food.
+     * @param item  The entity which is consuming the food.
      */
     @ReceiveEvent(components = ItemComponent.class, priority = EventPriority.PRIORITY_TRIVIAL)
     public void usedItem(FoodConsumedEvent event, EntityRef item) {
